// run-archive/scenario-2/1777109534353-scenario-2-deepseek-v4-pro-plan-flash-build-attempt-1/frontend/src/api/model/bookingResponse.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Restaurant Booking API
 * OpenAPI spec version: 1.0.0
 */

export interface BookingResponse {
  id: string;
  restaurantId: string;
  restaurantName: string;
  tableId: string;
  date: string;
  startTime: string;
  partySize: number;
  guestName: string;
  createdAt: string;
}